backend/requirements.txt
Django==4.2.16
python-dotenv==1.0.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
pillow==10.4.0
django-cors-headers==4.4.0
django-filter==23.1
celery==5.4.0
redis==5.1.1
requests==2.32.3